+++ internal/distro/rhel9/package_sets.go
@@ -468,137 +468,6 @@ func qcow2CommonPackageSet(t *imageType) rpmmd.PackageSet {
 	return ps
 }
 
-func vhdCommonPackageSet(t *imageType) rpmmd.PackageSet {
-	ps := rpmmd.PackageSet{
-		Include: []string{
-			// Defaults
-			"langpacks-en",
-			// From the lorax kickstart
-			"chrony",
-			"firewalld",
-			"WALinuxAgent",
-			"python3",
-			"net-tools",
-			"cloud-init",
-			"cloud-utils-growpart",
-			"gdisk",
-
-			// removed from defaults but required to boot in azure
-			"dhcp-client",
-		},
-		Exclude: []string{
-			"rng-tools",
-		},
-	}.Append(bootPackageSet(t)).Append(coreOsCommonPackageSet(t))
-
-	if t.arch.Name() == distro.X86_64ArchName {
-		ps = ps.Append(rpmmd.PackageSet{
-			Include: []string{
-				// packages below used to come from @core group and were not excluded
-				// they may not be needed at all, but kept them here to not need
-				// to exclude them instead in all other images
-				"iwl100-firmware",
-				"iwl105-firmware",
-				"iwl135-firmware",
-				"iwl1000-firmware",
-				"iwl2000-firmware",
-				"iwl2030-firmware",
-				"iwl3160-firmware",
-				"iwl5000-firmware",
-				"iwl5150-firmware",
-				"iwl6000g2a-firmware",
-				"iwl6050-firmware",
-				"iwl7260-firmware",
-			},
-		})
-	}
-
-	return ps
-}
-
-func azureRhuiCommonPackageSet(t *imageType) rpmmd.PackageSet {
-	return rpmmd.PackageSet{
-		Include: []string{
-			"@Server",
-			"bzip2",
-			"cloud-init",
-			"cloud-utils-growpart",
-			"dracut-config-generic",
-			"efibootmgr",
-			"gdisk",
-			"grub2-efi-x64",
-			"grub2-pc",
-			"hyperv-daemons",
-			"insights-client",
-			"kernel-core",
-			"kernel-modules",
-			"kernel",
-			"langpacks-en",
-			"lvm2",
-			"NetworkManager",
-			"NetworkManager-cloud-setup",
-			"nvme-cli",
-			"patch",
-			"rhc",
-			"rhui-azure-rhel9",
-			"rng-tools",
-			"selinux-policy-targeted",
-			"shim-x64",
-			"uuid",
-			"WALinuxAgent",
-			"yum-utils",
-		},
-		Exclude: []string{
-			"aic94xx-firmware",
-			"alsa-firmware",
-			"alsa-lib",
-			"alsa-sof-firmware",
-			"alsa-tools-firmware",
-			"biosdevname",
-			"bolt",
-			"buildah",
-			"cockpit-podman",
-			"containernetworking-plugins",
-			"dnf-plugin-spacewalk",
-			"dracut-config-rescue",
-			"glibc-all-langpacks",
-			"iprutils",
-			"ivtv-firmware",
-			"iwl100-firmware",
-			"iwl1000-firmware",
-			"iwl105-firmware",
-			"iwl135-firmware",
-			"iwl2000-firmware",
-			"iwl2030-firmware",
-			"iwl3160-firmware",
-			"iwl3945-firmware",
-			"iwl4965-firmware",
-			"iwl5000-firmware",
-			"iwl5150-firmware",
-			"iwl6000-firmware",
-			"iwl6000g2a-firmware",
-			"iwl6000g2b-firmware",
-			"iwl6050-firmware",
-			"iwl7260-firmware",
-			"libertas-sd8686-firmware",
-			"libertas-sd8787-firmware",
-			"libertas-usb8388-firmware",
-			"NetworkManager-config-server",
-			"plymouth",
-			"podman",
-			"python3-dnf-plugin-spacewalk",
-			"python3-hwdata",
-			"python3-rhnlib",
-			"rhn-check",
-			"rhn-client-tools",
-			"rhn-setup",
-			"rhnlib",
-			"rhnsd",
-			"usb_modeswitch",
-		},
-	}.Append(bootPackageSet(t))
-}
-
 func vmdkCommonPackageSet(t *imageType) rpmmd.PackageSet {
 	ps := rpmmd.PackageSet{
 		Include: []string{
